# backend/cmd/server/openapi.yaml
openapi: 3.0.3
info:
  title: Attic Asset Management API
  description: API for managing organizational assets, warranties, and attachments
  version: 1.0.0
  contact:
    name: Attic
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/me:
    get:
      tags: [Auth]
      summary: Get current user
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/categories:
    get:
      tags: [Categories]
      summary: List categories
      security:
        - sessionCookie: []
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
    post:
      tags: [Categories]
      summary: Create category
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
  /api/categories/{id}:
    get:
      tags: [Categories]
      summary: Get category by ID
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Category details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Categories]
      summary: Update category
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        '200':
          description: Category updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
    delete:
      tags: [Categories]
      summary: Delete category
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '204':
          description: Category deleted
  /api/locations:
    get:
      tags: [Locations]
      summary: List locations
      security:
        - sessionCookie: []
      responses:
        '200':
          description: List of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Location'
    post:
      tags: [Locations]
      summary: Create location
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationInput'
      responses:
        '201':
          description: Location created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
  /api/locations/{id}:
    get:
      tags: [Locations]
      summary: Get location by ID
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Location details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
    put:
      tags: [Locations]
      summary: Update location
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationInput'
      responses:
        '200':
          description: Location updated
    delete:
      tags: [Locations]
      summary: Delete location
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '204':
          description: Location deleted
  /api/conditions:
    get:
      tags: [Conditions]
      summary: List conditions
      security:
        - sessionCookie: []
      responses:
        '200':
          description: List of conditions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Condition'
    post:
      tags: [Conditions]
      summary: Create condition
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConditionInput'
      responses:
        '201':
          description: Condition created
  /api/conditions/{id}:
    get:
      tags: [Conditions]
      summary: Get condition by ID
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Condition details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Condition'
    put:
      tags: [Conditions]
      summary: Update condition
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConditionInput'
      responses:
        '200':
          description: Condition updated
    delete:
      tags: [Conditions]
      summary: Delete condition
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '204':
          description: Condition deleted
  /api/collections:
    get:
      tags: [Collections]
      summary: List shared collections
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Collection'
        '400':
          description: Invalid request
        '404':
          description: Collection not found in this workspace
    post:
      tags: [Collections]
      summary: Create collection
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionInput'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '400':
          description: Invalid request
        '404':
          description: Collection not found in this workspace
        '409':
          description: A collection with this name already exists (case insensitive)
  /api/collections/{id}:
    get:
      tags: [Collections]
      summary: Get collection
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '400':
          description: Invalid request
        '404':
          description: Collection not found in this workspace
    put:
      tags: [Collections]
      summary: Update collection
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionInput'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '400':
          description: Invalid request
        '404':
          description: Collection not found in this workspace
        '409':
          description: A collection with this name already exists (case insensitive)
    delete:
      tags: [Collections]
      summary: Delete collection without deleting assets
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '204':
          description: Success
        '400':
          description: Invalid request
        '404':
          description: Collection not found in this workspace
  /api/assets:
    get:
      tags: [Assets]
      summary: List assets
      security:
        - sessionCookie: []
      parameters:
        - name: q
          in: query
          description: Full-text search query
          schema:
            type: string
        - name: collection_id
          in: query
          description: Filter by a shared collection, combined with other filters
          schema:
            type: string
            format: uuid
        - name: category_id
          in: query
          description: Filter by a category UUID or use `uncategorized` for assets without a category
          schema:
            oneOf:
              - type: string
                format: uuid
              - type: string
                enum: [uncategorized]
        - name: location_id
          in: query
          schema:
            type: string
            format: uuid
        - name: condition_id
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Paginated list of assets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetList'
    post:
      tags: [Assets]
      summary: Create asset
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetInput'
      responses:
        '201':
          description: Asset created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
  /api/assets/{id}:
    get:
      tags: [Assets]
      summary: Get asset by ID
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Asset details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
    put:
      tags: [Assets]
      summary: Update asset
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetInput'
      responses:
        '200':
          description: Asset updated
    delete:
      tags: [Assets]
      summary: Delete asset
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '204':
          description: Asset deleted
  /api/assets/{id}/warranty:
    get:
      tags: [Warranties]
      summary: Get asset warranty
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Warranty details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Warranty'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Warranties]
      summary: Create warranty for asset
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WarrantyInput'
      responses:
        '201':
          description: Warranty created
    put:
      tags: [Warranties]
      summary: Update asset warranty
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WarrantyInput'
      responses:
        '200':
          description: Warranty updated
    delete:
      tags: [Warranties]
      summary: Delete asset warranty
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '204':
          description: Warranty deleted
  /api/warranties/expiring:
    get:
      tags: [Warranties]
      summary: List expiring warranties
      security:
        - sessionCookie: []
      parameters:
        - name: days
          in: query
          description: Number of days to look ahead
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: List of expiring warranties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Warranty'
  /api/assets/{id}/attachments:
    get:
      tags: [Attachments]
      summary: List asset attachments
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: List of attachments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Attachment'
    post:
      tags: [Attachments]
      summary: Upload attachment
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Attachment uploaded
  /api/attachments/{attachmentId}:
    get:
      tags: [Attachments]
      summary: Get attachment download URL
      security:
        - sessionCookie: []
      parameters:
        - name: attachmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Presigned download URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: uri
    delete:
      tags: [Attachments]
      summary: Delete attachment
      security:
        - sessionCookie: []
      parameters:
        - name: attachmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Attachment deleted
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: OIDC session cookie set after a successful login
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        parent_id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        color:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        parent_id:
          type: string
          format: uuid
        icon:
          type: string
        color:
          type: string
    Location:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        parent_id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    LocationInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        parent_id:
          type: string
          format: uuid
        icon:
          type: string
    Condition:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        display_order:
          type: integer
        color:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...